fix(migrations): widen dashboard_id to bigint for Grafana 12 compatibility
application_dashboards.dashboard_id and benchmarks.dashboard_id were still
integer (max ~2.1B) while Grafana 12 returns IDs like 1324859669958656.
Every insert to application_dashboards was failing, causing BenchmarkProcessorService
to find 0 dashboards and no benchmarks/SLOs to be created.

Mirrors the fix from ChangeGrafanaIdToBigint which widened grafana_dashboards.grafana_id
but missed these two columns.

- application-dashboard.entity.ts: type integer → bigint
- schema-sql.ts: dashboard_id bigint in application_dashboards + benchmarks tables
- schema-sql.ts: migrate_benchmark_from_mongodb p_dashboard_id bigint

+++ packages/shared/src/entities/application-dashboard.entity.ts
@@ -61,7 +61,7 @@ export class ApplicationDashboard {
   @Column({ name: 'dashboard_name', type: 'varchar', length: 255 })
   dashboardName!: string;
 
-  @Column({ name: 'dashboard_id', type: 'integer', nullable: true })
+  @Column({ name: 'dashboard_id', type: 'bigint', nullable: true })
   dashboardId?: number;
 
   @Column({ name: 'dashboard_uid', type: 'varchar', length: 100, nullable: true })
